fincal changes hopefully?

+++ ex00/src/main.cpp
@@ -19,7 +19,7 @@ int main() {
     std::cout << std::endl << "--------------------------------" << std::endl << std::endl;
 
     claptrap.setAttDmg(1);
-    for (uint8_t i = 0; i < 10; i++) {
+    for (int i = 0; i < 10; i++) {
         std::cout << "attack " << i << " -----" << std::endl;
         claptrap.attack("Springtrap");
         claptrap4.takeDamage(claptrap.getAttDmg());
